+++ packages/sources/src/components/pagination/pagination.args.ts
@@ -1,3 +1,4 @@
+import { HandlerFunction } from '@storybook/addon-actions';
 import { ArgTypes } from '../../stories-helpers';
 
 import { Pagination } from './pagination.models';
@@ -5,6 +6,8 @@ import { Pagination } from './pagination.models';
 export interface PaginationArgs {
   count: number;
   current: number;
+  label: string;
+  onSelectPage: HandlerFunction;
 }
 
 export const paginationArgTypes: ArgTypes<PaginationArgs> = {
@@ -17,12 +20,22 @@ export const paginationArgTypes: ArgTypes<PaginationArgs> = {
     control: {
       type: 'number'
     }
+  },
+  label: {
+    control: {
+      type: 'text'
+    }
+  },
+  onSelectPage: {
+    action: 'selectPage'
   }
 };
 
 export function paginationArgsMapper(a: PaginationArgs): Pagination {
   return {
     count: a.count,
-    current: a.current
+    current: a.current,
+    label: a.label,
+    onSelectPage: a.onSelectPage,
   };
 }