+++ packages/sources/src/components/pagination/pagination.models.ts
@@ -1,4 +1,6 @@
 export interface Pagination {
   count: number;
   current: number;
+  label: string;
+  onSelectPage: (e: CustomEvent<any>) => void;
 }